update code according to PR
