+++ LibOS/shim/test/fs/.gitignore
@@ -17,6 +17,7 @@
 /seek_tell
 /stat
 /truncate
+
 /copy_subdir_tmpfs
 /copy_tmpfs
 /delete_tmpfs
