+++ LibOS/shim/test/fs/common.h
@@ -44,5 +44,5 @@ void* alloc_buffer(size_t size);
 void fill_random(void* buffer, size_t size);
 
 void copy_data(int fi, int fo, const char* input_path, const char* output_path, size_t size);
-void copy_file_tmpfs(const char* input_path, const char* output_path);
+
 #endif
