+++ LibOS/shim/test/fs/copy_tmpfs.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "common_tmpfs.h"
 
 int main(int argc, char* argv[]) {
     if (argc < 4)
